backend/requirements.txt
addict
fastapi
flake8
flake8-docstrings
pyjwt
pymongo
python-multipart
typer
uvicorn

-e .